Remove option to return an error in MapFunc.
Cureent behavior when an error is returned is to return from StartBroadcast, passing the error to the caller. This means that if an error is returned the broadcast service will be stopped. Instead, the caller should handle the erros that could occur in MapFunc inside the implementation of the function, and panic if an error is big enough to stop the broadcast service.